requirements.txt
Flask==3.1.2
Requests==2.32.5
boto3
Flask-Mail
Pillow